+++ src/main/java/com/jcdecaux/datacorp/spark/enums/Storage.java
@@ -7,5 +7,7 @@ public enum Storage {
     CSV,
     EXCEL,
     PARQUET,
-    CASSANDRA
+    CASSANDRA,
+    DYNAMODB,
+    OTHER
 }